+++ src/store/store.utils.js
@@ -56,7 +56,7 @@ function getPrimaryKeyValue (store, value, isInsertOrMassiveUpdate) {
   }
   if (!store.primaryKey) {
     logger.tip(
-      'No primary key has been found, fallback to lunaris _id.',
+      'No primary key has been found in store "' + store.name + '", fallback to lunaris object attribute "_id".',
       'To declare a primary key, use the notation [\'<<int>>\'] in the map or add the \'primaryKey\' attribute in the store description.'
     );
     return value._id;