+++ tests/test.lunaris.collection.js
@@ -1561,6 +1561,168 @@ describe('lunaris internal collection', () => {
     });
   });
 
+  describe('setIndexIdValue()', () => {
+    it('should be defined', () => {
+      should(collection().setIndexIdValue).be.ok();
+    });
+
+    it('should not set index id value', () => {
+      var _collection = collection();
+      should(_collection.getIndexId()).eql([
+        [], []
+      ]);
+      _collection.setIndexIdValue(1, 2);
+      should(_collection.getIndexId()).eql([
+        [], []
+      ]);
+    });
+
+    it('should set index id value', () => {
+      var _collection = collection((item) => {
+        return item.id;
+      });
+      should(_collection.getIndexId()).eql([
+        [], []
+      ]);
+
+      _collection.add({ id : '_1', label : 'A' });
+
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+
+      _collection.setIndexIdValue(1, 2);
+      should(_collection.getIndexId()).eql([
+        [2], [1]
+      ]);
+    });
+
+    it('should set index id value = 0', () => {
+      var _collection = collection((item) => {
+        return item.id;
+      });
+      should(_collection.getIndexId()).eql([
+        [], []
+      ]);
+
+      _collection.add({ id : '_1', label : 'A' });
+
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+
+      _collection.setIndexIdValue(1, 0);
+      should(_collection.getIndexId()).eql([
+        [0], [1]
+      ]);
+    });
+
+    it('should not set index id value if key = null', () => {
+      var _collection = collection((item) => {
+        return item.id;
+      });
+      should(_collection.getIndexId()).eql([
+        [], []
+      ]);
+
+      _collection.add({ id : '_1', label : 'A' });
+
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+
+      _collection.setIndexIdValue(null, 0);
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+    });
+
+    it('should not set index id value if key = undefined', () => {
+      var _collection = collection((item) => {
+        return item.id;
+      });
+      should(_collection.getIndexId()).eql([
+        [], []
+      ]);
+
+      _collection.add({ id : '_1', label : 'A' });
+
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+
+      _collection.setIndexIdValue(undefined, 0);
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+    });
+
+    it('should not set index id value if value = undefined', () => {
+      var _collection = collection((item) => {
+        return item.id;
+      });
+      should(_collection.getIndexId()).eql([
+        [], []
+      ]);
+
+      _collection.add({ id : '_1', label : 'A' });
+
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+
+      _collection.setIndexIdValue(1, undefined);
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+    });
+
+    it('should not set index id value if value = null', () => {
+      var _collection = collection((item) => {
+        return item.id;
+      });
+      should(_collection.getIndexId()).eql([
+        [], []
+      ]);
+
+      _collection.add({ id : '_1', label : 'A' });
+
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+
+      _collection.setIndexIdValue(1, null);
+      should(_collection.getIndexId()).eql([
+        ['_1'], [1]
+      ]);
+    });
+
+    it('should set index id value with multiple in the collection', () => {
+      var _collection = collection((item) => {
+        return item.id;
+      });
+      should(_collection.getIndexId()).eql([
+        [], []
+      ]);
+
+      _collection.add({ id : '_1', label : 'A' });
+      _collection.add({ id : '_2', label : 'B' });
+      _collection.add({ id : '_3', label : 'C' });
+      _collection.add({ id : '_4', label : 'D' });
+
+      should(_collection.getIndexId()).eql([
+        ['_1', '_2', '_3', '_4'],
+        [1, 2, 3, 4]
+      ]);
+
+      _collection.setIndexIdValue(3, 33);
+      should(_collection.getIndexId()).eql([
+        ['_1', '_2', 33, '_4'],
+        [1, 2, 3, 4]
+      ]);
+    });
+  });
+
   describe('setIndexReferences()', () => {
     it('should be defined', () => {
       should(collection().setIndexReferences).be.a.Function();